refactor: extract getCredentialsWithPlaceholders to its own module
Move the helper out of EditCredentialsModal.tsx so the component file
exports only React components (fixes react-refresh/only-export-components
lint error introduced by exporting the helper for testing).
