+++ packages/frontend/src/components/UserSettings/MyWarehouseConnectionsPanel/EditCredentialsModal.tsx
@@ -1,5 +1,4 @@
 import {
-    WarehouseTypes,
     type UpsertUserWarehouseCredentials,
     type UserWarehouseCredentials,
 } from '@lightdash/common';
@@ -11,40 +10,8 @@ import { useUserWarehouseCredentialsUpdateMutation } from '../../../hooks/userWa
 import MantineModal, {
     type MantineModalProps,
 } from '../../common/MantineModal';
+import { getCredentialsWithPlaceholders } from './getCredentialsWithPlaceholders';
 import { WarehouseFormInputs } from './WarehouseFormInputs';
-
-export const getCredentialsWithPlaceholders = (
-    credentials: UserWarehouseCredentials['credentials'],
-): UpsertUserWarehouseCredentials['credentials'] => {
-    switch (credentials.type) {
-        case WarehouseTypes.REDSHIFT:
-        case WarehouseTypes.SNOWFLAKE:
-        case WarehouseTypes.POSTGRES:
-        case WarehouseTypes.TRINO:
-            return {
-                ...credentials,
-                password: '',
-            };
-        case WarehouseTypes.BIGQUERY:
-            return {
-                ...credentials,
-                keyfileContents: {},
-            };
-        case WarehouseTypes.DATABRICKS:
-            return {
-                ...credentials,
-                personalAccessToken: '',
-            };
-        case WarehouseTypes.ATHENA:
-            return {
-                ...credentials,
-                accessKeyId: '',
-                secretAccessKey: '',
-            };
-        default:
-            throw new Error(`Credential type not supported`);
-    }
-};
 
 const FORM_ID = 'edit-credentials-form';
